requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
Django==4.1.7
django-crispy-forms==2.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-tinymce==3.6.1
idna==3.4
Pillow==9.4.0
requests==2.28.2
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
